make DashTable component it's own file and add proptypes
